requirements.txt
# ═══════════════════════════════════════════════════
# DEPENDENCIAS DEL AGENTE DE RESEARCH
# ═══════════════════════════════════════════════════

# API y servidor web
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9        # Para form data en FastAPI

# LangChain — orquestación del agente
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0        # Proveedor Ollama
langchain-openai>=0.2.0        # TODO: Proveedor OpenAI
langchain-anthropic>=0.3.0     # TODO: Proveedor Anthropic
langchain-groq>=0.2.0          # TODO: Proveedor Groq

# Búsqueda web
ddgs>=1.0.0                    # Búsqueda DuckDuckGo sin API key (ex duckduckgo-search)
tavily-python>=0.5.0           # TODO: Búsqueda Tavily (requiere API key)

# Vector store
chromadb>=0.5.0

# Base de datos (TODO v2)
sqlalchemy>=2.0.0
aiosqlite>=0.20.0              # SQLite async

# Generación de PDF
weasyprint>=62.0

# Configuración
python-dotenv>=1.0.0

# Validación de datos
pydantic>=2.8.0
pydantic-settings>=2.4.0

# HTTP async (para health checks de proveedores)
httpx>=0.27.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0
